--- a/bin/test_e2e_compatibility
+++ b/bin/test_e2e_compatibility
@@ -66,7 +66,6 @@ do
             go test -v ./e2e/... -args \
             --provider.tequilapi-host=myst-provider-$i \
             --provider.tequilapi-port=4050 \
-            --consumer.tequilapi-host=myst-consumer-$j \
             --consumer.tequilapi-port=4050 \
             --geth.url=http://geth:8545 \
             --consumer.services=$consumerServices