--- a/ci/test/e2e.go
+++ b/ci/test/e2e.go
@@ -74,8 +74,8 @@ func TestE2EBasic() error {
 	composeFiles := []string{
 		"./docker-compose.e2e-basic.yml",
 	}
-	runner, _ := e2e.NewRunner(composeFiles, "node_e2e_basic_test", "openvpn,noop,wireguard")
-	// defer cleanup()
+	runner, cleanup := e2e.NewRunner(composeFiles, "node_e2e_basic_test", "openvpn,noop,wireguard")
+	defer cleanup()
 	if err := runner.Init(); err != nil {
 		return err
 	}