--- a/docker-compose.e2e-compatibility.yml
+++ b/docker-compose.e2e-compatibility.yml
@@ -90,7 +90,7 @@ services:
       - ./e2e/blockchain/keystore:/keystore
 
   accountant:
-    image: mysteriumnetwork/accountant:0.2
+    image: mysteriumnetwork/hermes:0.4.0
     environment:
       PORT: 8889
     expose: