--- a/docker-compose.e2e-traversal.yml
+++ b/docker-compose.e2e-traversal.yml
@@ -179,7 +179,7 @@ services:
         ipv4_address: 172.31.0.203
 
   accountant:
-    image: localhost:5000/accountant
+    image: mysteriumnetwork/hermes:0.4.0
     environment:
       PORT: 8889
     expose: