--- a/e2e/test_options.go
+++ b/e2e/test_options.go
@@ -42,13 +42,13 @@ var (
 
 // Consumer flags
 var (
-	consumerTequilapiHost = flag.String("consumer.tequilapi-host", "localhost", "Specify Tequilapi host for consumer")
 	consumerTequilapiPort = flag.Int("consumer.tequilapi-port", 4050, "Specify Tequilapi port for consumer")
 	consumerServices      = flag.String("consumer.services", "openvpn,noop,wireguard", "Comma separated list of services to try and use")
 )
 
-func newTequilapiConsumer() *tequilapi_client.Client {
-	return tequilapi_client.NewClient(*consumerTequilapiHost, *consumerTequilapiPort)
+func newTequilapiConsumer(serviceType string) *tequilapi_client.Client {
+	host := "myst-consumer-" + serviceType
+	return tequilapi_client.NewClient(host, *consumerTequilapiPort)
 }
 
 func newTequilapiProvider() *tequilapi_client.Client {