… cmd/di_desktop.go
… e2e/connection_test.go
… go.sum
… session/pingpong/accountant_promise_handler_test.go
… session/pingpong/factory.go
… session/pingpong/invoice_tracker.go
… session/pingpong/invoice_tracker_test.go