E2e now starts multiple consumers, these then connect to the provider simultaneously. This mimics the real world scenario more closely.

Had to refactor accountant promise handling flow, as this caused insane races. To fix it, moved all the accountant promise logic to a new component: `accountant_promise_handler.go`. It will handle the promises sequentially and avoid race conditions.


Time savings:
After
`mage teste2ebasic  32.50s user 17.48s system 24% cpu 3:20.12 total `
Before
`mage teste2ebasic  45.29s user 18.08s system 21% cpu 4:58.23 total`